refactor: share normalize function
